backend/requirements.txt
flask
flask-cors
pdfminer.six
python-dotenv
google-generativeai
pytesseract
Pillow
